requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
flask-session==0.5.0

# Core Dependencies
pandas==2.1.4
numpy==1.26.2

# Machine Learning & NLP
scikit-learn==1.3.2
spacy>=3.7.2

# Data Visualization
plotly==5.18.0

# API and Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-decouple==3.8

# AI/ML APIs
google-generativeai==0.8.3
